Move FakeTenant to utils so it's shared across backends
FakeTenant is now defined in django_tenants.utils instead of only in
the PostgreSQL backend, enabling both the PostgreSQL and MySQL backends
to use the same class. This ensures isinstance(connection.tenant, FakeTenant)
checks work correctly regardless of which backend is active, particularly
in django_tenants/template/loaders/cached.py.
